requirements.txt
celery==5.2.7
Django==3.2
django-cors-headers==3.10.1
django-filter==22.1
django-phonenumbers==1.0.1
django-phonenumber-field==5.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
psycopg2-binary
python-dotenv==0.21.0
python-dateutil==2.8.2
drf-yasg==1.21.4
Pillow==9.4.0
django-admin-interface==0.24.2
django-ckeditor==5.7.1
django-celery-results==2.0.0
fcm-django==1.0.8